+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Enhancements
 
 - ASE interface for DFTB+ calculations added
+- Add a new keyword 'freset_forces' to reset forces to zero after each step
 - init_velocities keyword is ignored if non-zero velocities are present
 - init_velocities can now be forced via the 'force' option